Fixed NullPointerException when the timer was already gone.
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@4629 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/main/java/org/red5/server/stream/PlayEngine.java
@@ -1295,7 +1295,8 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
 			case PipeConnectionEvent.PROVIDER_CONNECT_PUSH:
 				if (event.getProvider() != this) {
 					if (waiting) {
-						schedulingService.removeScheduledJob(waitLiveJob);
+						if(waitLiveJob != null)
+							schedulingService.removeScheduledJob(waitLiveJob);
 						waitLiveJob = null;
 						waiting = false;
 					}
